+++ GraphEditorKeyframeGeneratorModule.py
@@ -23,10 +23,10 @@ else:
 
 import bpy
 from collections import defaultdict
-from typing import List
+from typing import List, Tuple, Optional
 from .midi_data import MidiDataType
 from .NoteCollectionModule import NoteCollection, NoteCollectionMetaData, NoteCollectionOverlapStrategy, \
-    NoteCollectionFilter
+    NoteCollectionFilter, AnalyzedNote
 
 
 def existing_keyframe_map(keyframes):
@@ -51,18 +51,25 @@ def add_keyframe(keyframe_points, existing_keyframes, frame, keyframe_value, ove
     if overlap_mode == 'SKIP':
         if not existing_keyframes[frame]:
             add(frame)
+            return frame
     elif overlap_mode == 'PREVIOUS_FRAME':
         if not existing_keyframes[frame]:
             add(frame)
+            return frame
         elif not existing_keyframes[frame - 1]:
             add(frame - 1)
+            return frame - 1
     elif overlap_mode == 'NEXT_FRAME':
         if not existing_keyframes[frame]:
             add(frame)
+            return frame
         elif not existing_keyframes[frame + 1]:
             add(frame + 1)
+            return frame + 1
     else:  # replace
         add(frame)
+        return frame
+    return None
 
 
 note_property_definitions = \
@@ -101,16 +108,121 @@ class FcurveKeyframeGenerator:
         self.min_keyframe_value = min_keyframe_value
         self.max_keyframe_value = max_keyframe_value
         self.keyframe_generator_property = keyframe_generator_property
-        self.generate_at_note_end: bool = keyframe_generator_property.generate_at_note_end
+        self.keyframe_placement: str = keyframe_generator_property.keyframe_placement
         self.on_keyframe_overlap: str = keyframe_generator_property.on_keyframe_overlap
         self.operator = operator  # blender operator that can be used for report
         self.note_property: str = keyframe_generator_property.note_property
+        self.limit_transition_length: bool = keyframe_generator_property.limit_transition_length
+        self.transition_limit_frames: int = keyframe_generator_property.transition_limit_frames
+        self.transition_offset_frames: int = keyframe_generator_property.transition_offset_frames
+        self.transition_placement: str = keyframe_generator_property.transition_placement
+
+    def add_transition_frames(self, keyframe_points,
+                              transition_ranges: List[Tuple[Tuple[int, float], Tuple[int, float]]]):
+        """
+        :param keyframe_points: FCurveKeyframePoints
+        :param transition_ranges: starting transition ranges and values (first and last keyframes)
+        """
+
+        def keyframe_exists_in_range(range_exclusive: Tuple[Tuple[int, float], Tuple[int, float]]):
+            """
+            :param range_exclusive: existing first and last keyframes
+            :return: True if a keyframe exists between first and last (not including first and last)
+            """
+            for x in range(range_exclusive[0][0] + 1, range_exclusive[1][0]):
+                if x in existing_keyframes:
+                    return True
+            return False
+
+        if not self.limit_transition_length:
+            return
+
+        existing_keyframes = existing_keyframe_map(keyframe_points)
+        for transition_range in transition_ranges:
+            transition_frame_start = transition_range[0][0]
+            transition_frame_end = transition_range[1][0]
+            if (transition_frame_end - transition_frame_start) > self.transition_limit_frames and \
+                    not keyframe_exists_in_range(transition_range):
+                transition_frame_offset = min(
+                    transition_range[1][0] - transition_range[0][0] - self.transition_limit_frames,
+                    self.transition_offset_frames)
+                if self.transition_placement == "end":
+                    transition_frame_end = transition_frame_end - transition_frame_offset
+                    transition_frame_start = transition_frame_end - self.transition_limit_frames
+                else:
+                    transition_frame_start = transition_frame_start + transition_frame_offset
+                    transition_frame_end = transition_frame_start + self.transition_limit_frames
+                add_keyframe(keyframe_points, existing_keyframes, transition_frame_start, transition_range[0][1],
+                             "SKIP")
+                add_keyframe(keyframe_points, existing_keyframes, transition_frame_end, transition_range[1][1], "SKIP")
+
+        pass
+
+    def add_keyframes(self, notes: List[AnalyzedNote], keyframe_points,
+                      keyframe_value_from_note):
+        def add_keyframes(generate_at_note_end: bool, on_keyframe_overlap: str) -> List[Tuple[int, int, float]]:
+            """
+            :param generate_at_note_end: if True generate at the end of the note, else generate at the start
+            :param on_keyframe_overlap: behavior if existing keyframe on the same frame
+            :return: list of (note index, frame, keyframe value)
+            """
+            existing_keyframes = existing_keyframe_map(keyframe_points)
+            keyframe_data_added = []
+            for i in range(len(notes)):
+                note = notes[i]
+                keyframe_value = keyframe_value_from_note(note)
+                if keyframe_value is not None:
+                    frame = note.note_end_frame if generate_at_note_end else note.note_start_frame
+                    frame_added_to = add_keyframe(keyframe_points, existing_keyframes, frame, keyframe_value,
+                                                  on_keyframe_overlap)
+                    if frame_added_to is not None:
+                        keyframe_data_added.append((i, frame_added_to, keyframe_value))
+            return keyframe_data_added
+
+        def create_transition_ranges(start_list: List[Tuple[int, int, float]],
+                                     end_list: Optional[List[Tuple[int, int, float]]]) \
+                -> List[Tuple[Tuple[int, float], Tuple[int, float]]]:
+            """
+            :param start_list: list of (note index, frame, keyframe value)
+            :param end_list: optional list of (note index, frame, keyframe value)
+            :return: list of ((transition start frame, transition start value),
+            (transition end frame, transition end value)). Transitions between keyframes if only start_list is provided.
+            If start_list and end_list are provided, transitions are from the end of the notes (based on end_list) to
+            the start of the next note (based on start_list)
+            """
+            range_list = []
+            if end_list is None:
+                previous_frame_data = None
+                for frame_data in start_list:
+                    if previous_frame_data is not None and (previous_frame_data[0] + 1 == frame_data[0]):
+                        range_list.append((previous_frame_data[1:], frame_data[1:]))
+                    previous_frame_data = frame_data
+            else:
+                start_frame_data_index = 0
+                for end_frame_data in end_list:
+                    while start_frame_data_index < len(start_list) and start_list[start_frame_data_index][1] <= \
+                            end_frame_data[1]:
+                        start_frame_data_index += 1
+                    if start_frame_data_index < len(start_list) and start_list[start_frame_data_index][0] == \
+                            end_frame_data[0] + 1:
+                        range_list.append((end_frame_data[1:], start_list[start_frame_data_index][1:]))
+            return range_list
+
+        if self.keyframe_placement == "note_start":
+            self.add_transition_frames(keyframe_points,
+                                       create_transition_ranges(add_keyframes(False, self.on_keyframe_overlap), None))
+        elif self.keyframe_placement == "note_end":
+            self.add_transition_frames(keyframe_points,
+                                       create_transition_ranges(add_keyframes(True, self.on_keyframe_overlap), None))
+        elif self.keyframe_placement == "note_start_and_end":
+            self.add_transition_frames(keyframe_points,
+                                       create_transition_ranges(add_keyframes(False, self.on_keyframe_overlap),
+                                                                add_keyframes(True, "PREVIOUS_FRAME")))
 
     def generate_keyframes_from_range_list(self, value_range_list, notes):
         """
         :param value_range_list: list of values to generate keyframes based on
         :param notes: list of notes
-        :param value_from_note: function to get the value from the note
         """
         if len(value_range_list) == 0:
             return
@@ -122,12 +234,8 @@ class FcurveKeyframeGenerator:
             mapped_keyframe_value += interval
 
         keyframe_points = self.fcurve.keyframe_points
-        existing_keyframes = existing_keyframe_map(keyframe_points)
-        for note in notes:
-            value = value_from_analyzed_note(self.note_property, note)
-            if value in values_map:
-                frame = note.note_end_frame if self.generate_at_note_end else note.note_start_frame
-                add_keyframe(keyframe_points, existing_keyframes, frame, values_map[value], self.on_keyframe_overlap)
+        self.add_keyframes(notes, keyframe_points,
+                           lambda note: values_map.get(value_from_analyzed_note(self.note_property, note), None))
 
         # triggers Blender's re-calculation of keyframe control points.
         bpy.ops.transform.translate()
@@ -140,7 +248,6 @@ class FcurveKeyframeGenerator:
         """
 
         keyframe_points = self.fcurve.keyframe_points
-        existing_keyframes = existing_keyframe_map(keyframe_points)
         if value_range_low != value_range_high:
             slope = (self.max_keyframe_value - self.min_keyframe_value) / (value_range_high - value_range_low)
         else:
@@ -148,11 +255,8 @@ class FcurveKeyframeGenerator:
             # (can't map same value to two different numbers)
             slope = 0
         y_intercept = self.min_keyframe_value - slope * value_range_low
-        for note in notes:
-            value = value_from_analyzed_note(self.note_property, note)
-            keyframe_value = slope * value + y_intercept
-            frame = note.note_end_frame if self.generate_at_note_end else note.note_start_frame
-            add_keyframe(keyframe_points, existing_keyframes, frame, keyframe_value, self.on_keyframe_overlap)
+        self.add_keyframes(notes, keyframe_points,
+                           lambda note: value_from_analyzed_note(self.note_property, note) * slope + y_intercept)
 
         # triggers Blender's re-calculation of keyframe control points.
         bpy.ops.transform.translate()